Revised RQ based on group discussion
I use the word recurrence to describe repeats of the same request by the same person, which could indicate that the request was not satisfactorily addressed. But, I can see how that might be too vague. I am trying to be concise, perhaps at the expense of clarity, but I didn't want to say, for example, "resolution time" or anything else for which we don't actually have a direct measure in the data.